+++ spec/lib/parser_spec.rb
@@ -31,7 +31,39 @@ describe Surveyor::Parser do
   it "should return a survey object" do
     Surveyor::Parser.new.parse("survey 'hi' do\n end").is_a?(Survey).should be_true
   end
+  describe 'reference checking' do
+    it 'accepts Answer#reference_identifier via underscore or hash syntax' do
+      survey_text = <<END
+  survey "Numbers" do
+    section_one "One" do
+      q_1 "Select a number", :pick => :one
+      a "One", {:reference_identifier => "1"}
+      a_2 "Two"
+      a_3 "Three"
 
+      label_2 "One is the loneliest number..."
+      dependency :rule => "A"
+      condition_A :q_1, "==", {:answer_reference => "1"}
+
+      label_3 "Two can be as bad as one..."
+      dependency :rule => "A"
+      condition_A :q_1, "==", {:answer_reference => "2"}
+
+      label_4 "that you'll ever do"
+      dependency :rule => "A"
+      condition_A :q_1, "==", :a_1
+
+      label_5 "it's the loneliest number since the number one"
+      dependency :rule => "A"
+      condition_A :q_1, "==", :a_2
+    end
+  end
+END
+      survey = Surveyor::Parser.new.parse(survey_text)
+      survey.is_a?(Survey).should == true
+    end
+
+  end
 
   describe 'translations' do